Clean up code

+++ ros1/src/hardware/interface/march_hardware/include/march_hardware/motor_controller/odrive/odrive.h
@@ -87,7 +87,7 @@ private:
     ODriveAxisState getAxisState();
 
     int32_t getAbsolutePositionIU();
-    float getIncrementalPositionIU();
+    int32_t getIncrementalPositionIU();
     float getIncrementalVelocityIU();
 
     uint32_t getAxisError();
